회원 가입 테스트 코드 추가

+++ src/main/java/com/taekyeong/tkgram/repository/UserRepository.java
@@ -1,10 +1,11 @@
 package com.taekyeong.tkgram.repository;
 
-import com.taekyeong.tkgram.model.User;
+import com.taekyeong.tkgram.entity.User;
+import lombok.NoArgsConstructor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     List<User> findByEmail(String email);
 }
